+++ TabletopTweaks/Config/ModSettings.cs
@@ -10,6 +10,10 @@ namespace TabletopTweaks.Config {
         public static AddedContent AddedContent;
         public static Homebrew Homebrew;
         public static Blueprints Blueprints;
+        private static JsonSerializerSettings cachedSettings;
+        private static JsonSerializerSettings SerializerSettings {
+            get { if (cachedSettings == null) { cachedSettings = CreateSettings(); } return cachedSettings; }
+        }
 
         public static void LoadAllSettings() {
             LoadSettings("Fixes.json", ref Fixes);
@@ -18,6 +22,7 @@ namespace TabletopTweaks.Config {
             LoadSettings("Blueprints.json", ref Blueprints);
         }
         private static void LoadSettings<T>(string fileName, ref T setting) where T : IUpdatableSettings {
+            JsonSerializer serializer = JsonSerializer.Create(SerializerSettings);
             var assembly = Assembly.GetExecutingAssembly();
             string userConfigFolder = ModEntry.Path + "UserSettings";
             Directory.CreateDirectory(userConfigFolder);
@@ -25,13 +30,15 @@ namespace TabletopTweaks.Config {
             var userPath = $"{userConfigFolder}{Path.DirectorySeparatorChar}{fileName}";
 
             using (Stream stream = assembly.GetManifestResourceStream(resourcePath))
-            using (StreamReader reader = new StreamReader(stream)) {
-                setting = JsonConvert.DeserializeObject<T>(reader.ReadToEnd());
+            using (StreamReader streamReader = new StreamReader(stream))
+            using (JsonReader jsonReader = new JsonTextReader(streamReader)) {
+                setting = serializer.Deserialize<T>(jsonReader);
             }
             if (File.Exists(userPath)) {
-                using (StreamReader reader = File.OpenText(userPath)) {
+                using (StreamReader streamReader = File.OpenText(userPath))
+                using (JsonReader jsonReader = new JsonTextReader(streamReader)) {
                     try {
-                        T userSettings = JsonConvert.DeserializeObject<T>(reader.ReadToEnd());
+                        T userSettings = serializer.Deserialize<T>(jsonReader);
                         setting.OverrideSettings(userSettings);
                     } catch {
                         Main.Error("Failed to load user settings. Settings will be rebuilt.");
@@ -39,7 +46,26 @@ namespace TabletopTweaks.Config {
                     }
                 }
             }
-            File.WriteAllText(userPath, JsonConvert.SerializeObject(setting, Formatting.Indented));
+
+            using (StreamWriter streamWriter = new StreamWriter(userPath))
+            using (JsonWriter jsonWriter = new JsonTextWriter(streamWriter)) {
+                serializer.Serialize(jsonWriter, setting);
+            }
+        }
+        private static JsonSerializerSettings CreateSettings() {
+            var RefJsonSerializerSettings = new JsonSerializerSettings {
+                CheckAdditionalContent = false,
+                ConstructorHandling = ConstructorHandling.AllowNonPublicDefaultConstructor,
+                DefaultValueHandling = DefaultValueHandling.Include,
+                FloatParseHandling = FloatParseHandling.Double,
+                Formatting = Formatting.Indented,
+                MetadataPropertyHandling = MetadataPropertyHandling.ReadAhead,
+                MissingMemberHandling = MissingMemberHandling.Ignore,
+                NullValueHandling = NullValueHandling.Ignore,
+                ObjectCreationHandling = ObjectCreationHandling.Replace,
+                StringEscapeHandling = StringEscapeHandling.Default,
+            };
+            return RefJsonSerializerSettings;
         }
     }
 }